Seperate the server pool from the client logic
Add the Pool class which just deals with the pool of available servers and
all of the locking.
Use an instance of Pool in the MistClient class.
Allow the user to pass a server hostname directly to the client (so that we can
route delete actions to the server that actually has the container!)